Implement limited video support
MP4 videos are now included in the gallery.  The support is rather
limited: a) they are always appended at the end after the images; b) their
aspect must be 16:9 to tile properly (this is to avoid probing the video
metadata).

Video support required significant rework of the logic of the single-image
browsing mode.