Remove srcset functionality
SRCSET for thumbnails saved some bandwidth on load, but at this point
this is really too much of an optimization and it was annoying to keep
creating two versions of the thumbnail.  From now on there should only be
'thumb.jpg'.  It is show at 128x128 pixels, so if you want to support DPR
of 2, just supply 256x256 px images.